refactor(webui): move hash-scroll subscription from site-header into router init
Hash-after-navigation scrolling was tied to a useEffect in SiteHeader,
giving it a misleading owner — the scroll concern isn't header-scoped.
Bind it once at router init via `router.subscribe('onResolved', ...)`
so the layout tree stays free of this side effect.
